+++ src/types/WH.ts
@@ -2,6 +2,6 @@ import {Integer} from './Integer'
 
 /** Width and height in pixels. */
 export interface WH {
-  readonly w: Integer
-  readonly h: Integer
+  w: Integer
+  h: Integer
 }