+++ src/types/XY.ts
@@ -2,6 +2,6 @@ import {Integer} from './Integer'
 
 /** Cartesian coordinates in pixels with a top-left origin. */
 export interface XY {
-  readonly x: Integer
-  readonly y: Integer
+  x: Integer
+  y: Integer
 }